\name{stepRegressionC-package}
\alias{stepRegressionC-package}
\alias{stepRegressionC}
\docType{package}
\title{
\packageTitle{stepRegressionC}
}
\description{
\packageDescription{stepRegressionC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stepRegressionC}
\packageIndices{stepRegressionC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stepRegressionC}

Maintainer: \packageMaintainer{stepRegressionC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
